+++ lang/en/charon.php
@@ -9,6 +9,8 @@ $string['pluginadministration'] = 'Charon administration';
 $string['modulenameplural'] = 'Charons';
 $string['capella:addinstance'] = 'Add a new Charon';
 $string['plagiarism_no_connection'] = 'Failed to connect to Plagiarism';
+$string['gradenoun'] = 'Grade';
+$string['hidden'] = 'Hidden';
 
 /*------------------------*\
    Course settings form
@@ -213,8 +215,14 @@ $string['reset_to_templates'] = 'Reset templates';
 $string['feedback-text-single-submission'] = 'Feedback for this submission';
 $string['no-feedback-info'] = 'When a teacher adds feedback for the submission, it will be visible here.';
 $string['feedback-text-all-submissions'] = 'Feedback for all submissions for this charon';
-$string['showing-mail'] = 'Showing mail';
+$string['showing-email'] = 'Showing e-mail';
 $string['showing-table'] = 'Showing table';
+
+/*------------------------*\
+   Labels
+\*------------------------*/
+
+$string['label_name_new_review_comment'] = 'New';
 
 /*------------------------*\
    Messages