+++ plugin/app/Repositories/StudentsRepository.php
@@ -2,7 +2,9 @@
 
 namespace TTU\Charon\Repositories;
 
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 use Zeizig\Moodle\Models\Course;
 use Zeizig\Moodle\Models\GradeGrade;
 use Zeizig\Moodle\Models\GradeItem;
@@ -16,27 +18,51 @@ use Zeizig\Moodle\Models\User;
 class StudentsRepository
 {
     /**
+     * Search users by given course id.
+     *
      * @param integer $courseId
-     * @param string $keyword
+     * @param ?string $keyword
+     * @param string[] $roleNames
+     *
+     * @return Collection
      */
-    public function searchStudentsByCourseAndKeyword($courseId, $keyword)
-    {
+    public function searchUsersByCourseKeywordAndRole(
+        int $courseId,
+        ?string $keyword = null,
+        array $roleNames = []
+    ): Collection {
+
         $keyword = '%' . strtolower($keyword) . '%';
 
-        return DB::table('role_assignments')
+        $query = DB::table('role_assignments')
+            ->join('role', 'role_assignments.roleid', '=', 'role.id')
             ->join('user', 'role_assignments.userid', '=', 'user.id')
             ->join('context', 'role_assignments.contextid', '=', 'context.id')
             ->where('context.contextlevel', CONTEXT_COURSE)
-            ->where('context.instanceid', $courseId)
-            ->where(function ($query) use ($keyword) {
-                $query->whereRaw('LOWER(idnumber) like ?', [$keyword])
+            ->where('context.instanceid', $courseId);
+
+        if (!empty($roleNames)) {
+            $query->whereIn('role.shortname', $roleNames);
+        }
+
+        if ($keyword !== null) {
+            $query->where(function ($queryRaw) use ($keyword) {
+                $queryRaw->whereRaw('LOWER(idnumber) like ?', [$keyword])
                     ->orWhereRaw('LOWER(username) like ?', [$keyword])
                     ->orWhereRaw('LOWER(firstname) like ?', [$keyword])
                     ->orWhereRaw('LOWER(lastname) like ?', [$keyword])
                     ->orWhereRaw("CONCAT(LOWER(firstname), ' ', LOWER(lastname)) like ?", [$keyword]);
-            })
-            ->select('user.id', 'idnumber', 'username', 'firstname', 'lastname', DB::raw("CONCAT(firstname, ' ',lastname, ' (',username, ')') AS fullname"))
-            ->get();
+            });
+        }
+
+        return $query->select(
+            'user.id',
+            'idnumber',
+            'username',
+            'firstname',
+            'lastname',
+            DB::raw("CONCAT(firstname, ' ', lastname, ' (', username, ')') AS fullname")
+        )->get();
     }
 
     /**
@@ -66,5 +92,20 @@ class StudentsRepository
         } else {
             return 0;
         }
+    }
+
+    public function getAllEnrolled(int $courseId)
+    {
+        $context = \context_course::instance($courseId);
+        $users = get_enrolled_users($context);
+        return array_map(function ($user) {
+            $updatedUser = new \stdClass();
+            $updatedUser->id = $user->id;
+            $updatedUser->username = $user->username;
+            $updatedUser->firstname = $user->firstname;
+            $updatedUser->lastname = $user->lastname;
+
+            return $updatedUser;
+        }, $users);
     }
 }