+++ plugin/resources/assets/js/components/partials/FilesWithReviewComments.vue
@@ -5,7 +5,20 @@
                 <span class="header">
                     {{ file.path }}
                 </span>
-                <v-card v-for="reviewComment in file.reviewComments" :key="reviewComment.id" class="review-comment">
+                <span class="review-comment-submission">
+                  Submission: {{ file.submissionCreation }}
+                </span>
+                <a v-if="view==='student' && isSubmissionButtonAllowed(file.submissionId)" class="button is-link review-comment-submission-button" @click="changeAssignmentSubmissionUrl(file.submissionId);">
+                  Go to submission!
+                </a>
+                <v-btn v-if="view==='teacher' && isSubmissionButtonAllowed(file.submissionId)" class="ma-2" tile outlined color="primary" @click="changePopupSubmissionUrl(file.submissionId);">
+                  Go to submission!
+                </v-btn>
+                <v-card v-for="reviewComment in file.reviewComments"
+                        :key="reviewComment.id"
+                        class="review-comment"
+                        :class="{ 'notify-review-comment': notify(reviewComment) }"
+                >
                     <div class="review-comment-heading">
                         <div class="review-comment-heading-info">
                             <span class="review-comment-author">
@@ -14,16 +27,24 @@
                             <span class="review-comment-date">
                                 Comment created: {{ reviewComment.commentCreation }}
                             </span>
-                            <span class="review-comment-submission">
-                                Submission: {{ file.submissionCreation }}
-                            </span>
-                            <v-btn v-if="view==='teacher'"
-                                   icon
-                                   class="remove-button"
-                                   @click="deleteReviewComment(reviewComment.id, file.charonId)"
-                            >
-                                <img src="/mod/charon/pix/bin.png" alt="delete" width="24px">
-                            </v-btn>
+
+                            <div class="review-comment-heading-extra">
+
+                                <label-component
+                                    v-if="notify(reviewComment)"
+                                    :name="translate('labelNameNewReviewComment')"
+                                    :background-color="'red'"
+                                />
+
+                                <v-btn
+                                    v-if="view==='teacher'"
+                                    icon
+                                    class="remove-button"
+                                    @click="deleteReviewComment(reviewComment.id, file.charonId)"
+                                >
+                                    <img src="/mod/charon/pix/bin.png" alt="delete" width="24px">
+                                </v-btn>
+                            </div>
                         </div>
                     </div>
                     <div class="review-comment-body">
@@ -40,15 +61,32 @@
 <script>
 
 import {ReviewComment} from "../../api";
+import LabelComponent from "./LabelComponent";
+import {Translate} from "../../mixins";
 
 export default {
     name: "FilesWithReviewComments",
+    mixins: [Translate],
+    components: {LabelComponent},
     props: {
         filesWithReviewComments: { required: true },
-        view: { required: true }
+        view: { required: true },
+        openSubmissionId: { required:true }
     },
 
-    methods: {
+  methods: {
+        isSubmissionButtonAllowed(submissionId) {
+          return submissionId !== this.openSubmissionId;
+        },
+
+        changeAssignmentSubmissionUrl(submissionId) {
+          VueEvent.$emit('change-assignment-submission-url', submissionId);
+        },
+
+        changePopupSubmissionUrl(submissionId) {
+          VueEvent.$emit('change-popup-submission-url', submissionId);
+        },
+
         deleteReviewComment(reviewCommentId, charonId) {
             if (reviewCommentId === null) {
                 return;
@@ -59,11 +97,20 @@ export default {
                 VueEvent.$emit('show-notification', 'Review comment deleted!')
             });
         },
+
+        notify(reviewComment) {
+            return this.view === "student" && reviewComment.notify === 1;
+        }
     }
 }
 </script>
 
 <style scoped>
+
+.review-comment-submission-button {
+  margin-top: 0.5em;
+  margin-left: 1em;
+}
 
 .header {
     text-align: center;
@@ -117,7 +164,7 @@ p {
     margin-top: 1em;
 }
 
-.remove-button {
+.review-comment-heading-extra {
     float: right;
 }
 
@@ -126,6 +173,10 @@ p {
     padding-right: 0.5em;
     font-weight: normal;
     margin-left: 1em;
+}
+
+.notify-review-comment {
+    background-color: #f0f0ff!important;
 }
 
 </style>