+++ plugin/resources/assets/js/components/partials/index.js
@@ -4,11 +4,12 @@ import Datepicker from './Datepicker.vue'
 import FilesComponent from './FilesComponent.vue'
 import FilesComponentWithoutTree from './FilesComponentWithoutTree.vue'
 import FilesWithReviewComments from "./FilesWithReviewComments";
+import LabelComponent from "./LabelComponent";
 import SubmissionTableComponent from "./SubmissionTableComponent";
 import ToggleButton from "./ToggleButton";
 import MinMaxSlider from "./MinMaxSlider";
 
 export {
     CharonTab, CharonTabs, Datepicker, FilesComponent, FilesComponentWithoutTree,
-    FilesWithReviewComments, SubmissionTableComponent, ToggleButton, MinMaxSlider
+    FilesWithReviewComments, SubmissionTableComponent, ToggleButton, MinMaxSlider, LabelComponent
 }