+++ CHANGELOG.md
@@ -6,7 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.3] - 2022-05-11
+
 ### Added
+- #641 New grading method 'prefer_best_each_grade', which lets students build up points for each grade - as long as grades' individual percentage stays the same or improves
+- #711 Warn user of the possibly long waiting time for recalculating grades if 'Recalculate' is checked - and Charon's grading method is 'prefer_best_each_test_grade'
+- #721 Allow recalculation of Charon grades without any change in tests' max points or deadlines
+- #846 If the user is not a TalTech student, he will get email on to his own added email
+
+### Fixed
+- #702 Fix grading type saving in charon settings
+
+## [1.6.2] - 2022-04-14
+
+### Added
+- #709 Added direct links to submissions, they can be opened from notifications and submissions modal
+- #736 Notifying review comments have a slightly different color and a 'New'/'Uus' label for students and all of charons' review comments are ordered by submission creation date time (descending)
+- #734 Students distribution data query in popup now on demand, using button
+
+## [1.6.1] - 2022-03-11
+
+### Added
+- #732 Group submissions in popup will have common line height also in case of a wrap.
+- #751 Add an option to access query logs and turn on query logging for specific users
 
 - #806 Added tables and graph for student plagiarism history under Student Overview
 - #818 Integrated history changes from plagiarism
@@ -20,13 +42,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - #781 Remove code span padding-left from assignment.css
 - #795 Restored Mobile-ID component
-
+- #745 Rounded student's percentage in Charon assignment grademap table up to 2 decimal points
 
 ## [1.6.0] - 2022-02-10
 
 ### Added
 - #691 ReviewComment count for each submission added to "Latest submissions" and submission search by student in Popup
-- #676 Student has an option to toggle current submission comments and all comments for this charon. Same for the teacher.
+- #676 Student has an option to toggle current submission comments and all comments for this charon. Same for the teacher
 - #706 Adjust fetching of the submission file in student view
 - #677 Submission feedback is now available in table view when clicking toggle, like from student view
 - #723 Update dependencies so security check does not fail & Laravel
@@ -62,7 +84,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.5.5] - 2021-10-22
 
 ### Added
-- #589 When a teacher adds a new review comment about a student's submission, an eye symbol on the same submission will turn 
+- #589 When a teacher adds a new review comment about a student's submission, an eye symbol on the same submission will turn
   red on the student's view. When submission is opened by the student, the red color will be removed again.
 - #678 Added logging when an exception is thrown when charon is being deleted by moodle
 - #681 updated composer/composer to 2.1.9
@@ -208,7 +230,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.4] - 2021-03-16
 
 ### Fixed
-- \#399 Triggering result recalculation previously reset the persistent grade result value to 0 
+- \#399 Triggering result recalculation previously reset the persistent grade result value to 0
 
 ## [1.1.3] - 2021-03-16
 
@@ -243,7 +265,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - \#391 Labs can now be named
 
 ### Changed
-- \#394 Submission result colors in Student view are less random and reflect score % 
+- \#394 Submission result colors in Student view are less random and reflect score %
 
 ### Fixed
 - \#378 Total grade calculation in Grading view no longer shows constant 0 when forward slash is present in Grade idNumber